config/x86: Add EROFS_FS and CACHEFILES for image-granularity acceleration
In the container-based deployment cases, CUBE believes that container image
distribution can be dramatically optimized, which relies on understanding
and reorganizing the container image format, and that tech needs in this
direction require a bit of kernel-space help, starting with kernel features
that are already upstream.

In terms of specific CONFIG choices, tk5 will use the same values as the
default configuration of fedora40-config-6.10.8-200.fc40.x86_64.

Requested-by: Changpeng Liu <changpeliu@tencent.com>
Signed-off-by: Like Xu <likexu@tencent.com>